Move logger initialization to initialize module
Initializing the logger module earlier makes the tool work correctly
when trying to find the /etc/system-release file, and if it is not
there, throw a critical error to the user before anything else happens.

If we don't move the module to this point, the first /etc/system-releaes
check will silently fail and only the next call will throw the
exception, because at this point, we already have the logger
initialized.

To prevent further executions in the system, it is more reliable to
initialize the logger module earlier.
