+++ convert2rhel/main.py
@@ -19,7 +19,6 @@ __metaclass__ = type
 
 import logging
 import os
-import sys
 
 from convert2rhel import actions, applock, backup, breadcrumbs, checks, exceptions, grub
 from convert2rhel import logger as logger_module
@@ -44,40 +43,6 @@ class ConversionPhase:
     POST_PONR_CHANGES = 4
 
 
-def initialize_logger():
-    """
-    Entrypoint function that aggregates other calls for initialization logic
-    and setup for logger handlers that do not require root.
-    """
-
-    return logger_module.setup_logger_handler()
-
-
-def initialize_file_logging(log_name, log_dir):
-    """
-    Archive existing file logs and setup all logging handlers that require
-    root, like FileHandlers.
-
-    This function should be called after
-    :func:`~convert2rhel.main.initialize_logger`.
-
-    .. warning::
-        Setting log_dir underneath a world-writable directory (including
-        letting it be user settable) is insecure.  We will need to write
-        some checks for all calls to `os.makedirs()` if we allow changing
-        log_dir.
-
-    :param str log_name: Name of the logfile to archive and log to
-    :param str log_dir: Directory where logfiles are stored
-    """
-    try:
-        logger_module.archive_old_logger_files(log_name, log_dir)
-    except (IOError, OSError) as e:
-        loggerinst.warning("Unable to archive previous log: %s" % e)
-
-    logger_module.add_file_handler(log_name, log_dir)
-
-
 def main():
     """
     Wrapper around the main entrypoint.
@@ -86,9 +51,6 @@ def main():
     conversion process itself, then calls main_locked(), protected by
     the application lock, to do the conversion process.
     """
-
-    # initialize logging
-    initialize_logger()
 
     # handle command line arguments
     toolopts.CLI()
@@ -110,10 +72,6 @@ def main_locked():
 
     pre_conversion_results = None
     process_phase = ConversionPhase.POST_CLI
-
-    # since we now have root, we can add the FileLogging
-    # and also archive previous logs
-    initialize_file_logging("convert2rhel.log", logger_module.LOG_DIR)
 
     try:
         perform_boilerplate()
