+++ convert2rhel/unit_tests/initialize_test.py
@@ -17,11 +17,16 @@
 
 __metaclass__ = type
 
-import os
-
 import pytest
+import six
 
-from convert2rhel import applock, initialize, main
+from convert2rhel import applock, initialize
+from convert2rhel import logger as logger_module
+from convert2rhel import main
+
+
+six.add_move(six.MovedModule("mock", "mock", "unittest.mock"))
+from six.moves import mock
 
 
 @pytest.mark.parametrize(
@@ -32,6 +37,49 @@ from convert2rhel import applock, initialize, main
     ),
 )
 def test_run(monkeypatch, exit_code, tmp_path):
+    monkeypatch.setattr(logger_module, "LOG_DIR", str(tmp_path))
     monkeypatch.setattr(main, "main", value=lambda: exit_code)
     monkeypatch.setattr(applock, "_DEFAULT_LOCK_DIR", str(tmp_path))
     assert initialize.run() == exit_code
+
+
+def test_initialize_logger(monkeypatch):
+    setup_logger_handler_mock = mock.Mock()
+
+    monkeypatch.setattr(
+        logger_module,
+        "setup_logger_handler",
+        value=setup_logger_handler_mock,
+    )
+
+    initialize.initialize_logger()
+    setup_logger_handler_mock.assert_called_once()
+
+
+@pytest.mark.parametrize(("exception_type", "exception"), ((IOError, True), (OSError, True), (None, False)))
+def test_initialize_file_logging(exception_type, exception, monkeypatch, caplog):
+    add_file_handler_mock = mock.Mock()
+    archive_old_logger_files_mock = mock.Mock()
+
+    if exception:
+        archive_old_logger_files_mock.side_effect = exception_type
+
+    monkeypatch.setattr(
+        logger_module,
+        "add_file_handler",
+        value=add_file_handler_mock,
+    )
+    monkeypatch.setattr(
+        logger_module,
+        "archive_old_logger_files",
+        value=archive_old_logger_files_mock,
+    )
+
+    initialize.initialize_file_logging("convert2rhel.log", "/tmp")
+
+    if exception:
+        assert caplog.records[-1].levelname == "WARNING"
+        assert "Unable to archive previous log:" in caplog.records[-1].message
+
+    add_file_handler_mock.assert_called_once()
+    archive_old_logger_files_mock.assert_called_once()
