+++ convert2rhel/unit_tests/main_test.py
@@ -29,9 +29,22 @@ from convert2rhel import backup
 six.add_move(six.MovedModule("mock", "mock", "unittest.mock"))
 from six.moves import mock
 
-from convert2rhel import actions, applock, checks, exceptions, grub
-from convert2rhel import logger as logger_module
-from convert2rhel import main, pkghandler, pkgmanager, redhatrelease, repo, subscription, toolopts, unit_tests, utils
+from convert2rhel import (
+    actions,
+    applock,
+    checks,
+    exceptions,
+    grub,
+    main,
+    pkghandler,
+    pkgmanager,
+    redhatrelease,
+    repo,
+    subscription,
+    toolopts,
+    unit_tests,
+    utils,
+)
 from convert2rhel.actions import report
 from convert2rhel.breadcrumbs import breadcrumbs
 from convert2rhel.systeminfo import system_info
@@ -40,8 +53,6 @@ from convert2rhel.unit_tests import (
     CLIMocked,
     CollectEarlyDataMocked,
     FinishCollectionMocked,
-    InitializeFileLoggingMocked,
-    InitializeLoggerMocked,
     MainLockedMocked,
     PrintDataCollectionMocked,
     PrintSystemInformationMocked,
@@ -121,48 +132,6 @@ class TestRollbackChanges:
             main.rollback_changes()
 
 
-def test_initialize_logger(monkeypatch):
-    setup_logger_handler_mock = mock.Mock()
-
-    monkeypatch.setattr(
-        logger_module,
-        "setup_logger_handler",
-        value=setup_logger_handler_mock,
-    )
-
-    main.initialize_logger()
-    setup_logger_handler_mock.assert_called_once()
-
-
-@pytest.mark.parametrize(("exception_type", "exception"), ((IOError, True), (OSError, True), (None, False)))
-def test_initialize_file_logging(exception_type, exception, monkeypatch, caplog):
-    add_file_handler_mock = mock.Mock()
-    archive_old_logger_files_mock = mock.Mock()
-
-    if exception:
-        archive_old_logger_files_mock.side_effect = exception_type
-
-    monkeypatch.setattr(
-        logger_module,
-        "add_file_handler",
-        value=add_file_handler_mock,
-    )
-    monkeypatch.setattr(
-        logger_module,
-        "archive_old_logger_files",
-        value=archive_old_logger_files_mock,
-    )
-
-    main.initialize_file_logging("convert2rhel.log", "/tmp")
-
-    if exception:
-        assert caplog.records[-1].levelname == "WARNING"
-        assert "Unable to archive previous log:" in caplog.records[-1].message
-
-    add_file_handler_mock.assert_called_once()
-    archive_old_logger_files_mock.assert_called_once()
-
-
 class TestShowEula:
     eula_dir = os.path.realpath(os.path.join(os.path.dirname(__file__), "..", "data", "version-independent"))
 
@@ -216,8 +185,6 @@ def test_help_exit(monkeypatch, tmp_path):
     monkeypatch.setattr(applock, "_DEFAULT_LOCK_DIR", str(tmp_path))
     monkeypatch.setattr(sys, "argv", ["convert2rhel", "--help"])
     monkeypatch.setattr(utils, "require_root", RequireRootMocked())
-    monkeypatch.setattr(main, "initialize_logger", InitializeLoggerMocked())
-    monkeypatch.setattr(main, "initialize_file_logging", InitializeFileLoggingMocked())
     monkeypatch.setattr(main, "main_locked", MainLockedMocked())
 
     with pytest.raises(SystemExit):
@@ -228,8 +195,6 @@ def test_help_exit(monkeypatch, tmp_path):
 
 def test_main(monkeypatch, tmp_path):
     require_root_mock = mock.Mock()
-    initialize_logger_mock = mock.Mock()
-    initialize_file_logging_mock = mock.Mock()
     toolopts_cli_mock = mock.Mock()
     show_eula_mock = mock.Mock()
     print_data_collection_mock = mock.Mock()
@@ -255,8 +220,6 @@ def test_main(monkeypatch, tmp_path):
 
     monkeypatch.setattr(applock, "_DEFAULT_LOCK_DIR", str(tmp_path))
     monkeypatch.setattr(utils, "require_root", require_root_mock)
-    monkeypatch.setattr(main, "initialize_logger", initialize_logger_mock)
-    monkeypatch.setattr(main, "initialize_file_logging", initialize_file_logging_mock)
     monkeypatch.setattr(toolopts, "CLI", toolopts_cli_mock)
     monkeypatch.setattr(main, "show_eula", show_eula_mock)
     monkeypatch.setattr(breadcrumbs, "print_data_collection", print_data_collection_mock)
@@ -282,7 +245,6 @@ def test_main(monkeypatch, tmp_path):
 
     assert main.main() == 0
     assert require_root_mock.call_count == 1
-    assert initialize_logger_mock.call_count == 1
     assert toolopts_cli_mock.call_count == 1
     assert show_eula_mock.call_count == 1
     assert print_data_collection_mock.call_count == 1
@@ -308,8 +270,6 @@ def test_main(monkeypatch, tmp_path):
 class TestRollbackFromMain:
     def test_main_rollback_post_cli_phase(self, monkeypatch, caplog, tmp_path):
         require_root_mock = mock.Mock()
-        initialize_logger_mock = mock.Mock()
-        initialize_file_logging_mock = mock.Mock()
         toolopts_cli_mock = mock.Mock()
         show_eula_mock = mock.Mock(side_effect=Exception)
 
@@ -318,15 +278,12 @@ class TestRollbackFromMain:
 
         monkeypatch.setattr(applock, "_DEFAULT_LOCK_DIR", str(tmp_path))
         monkeypatch.setattr(utils, "require_root", require_root_mock)
-        monkeypatch.setattr(main, "initialize_logger", initialize_logger_mock)
-        monkeypatch.setattr(main, "initialize_file_logging", initialize_file_logging_mock)
         monkeypatch.setattr(toolopts, "CLI", toolopts_cli_mock)
         monkeypatch.setattr(main, "show_eula", show_eula_mock)
         monkeypatch.setattr(breadcrumbs, "finish_collection", finish_collection_mock)
 
         assert main.main() == 1
         assert require_root_mock.call_count == 1
-        assert initialize_logger_mock.call_count == 1
         assert toolopts_cli_mock.call_count == 1
         assert show_eula_mock.call_count == 1
         assert finish_collection_mock.call_count == 1
@@ -335,8 +292,6 @@ class TestRollbackFromMain:
     def test_main_traceback_in_clear_versionlock(self, caplog, monkeypatch, tmp_path):
         monkeypatch.setattr(applock, "_DEFAULT_LOCK_DIR", str(tmp_path))
         monkeypatch.setattr(utils, "require_root", RequireRootMocked())
-        monkeypatch.setattr(main, "initialize_logger", InitializeLoggerMocked())
-        monkeypatch.setattr(main, "initialize_file_logging", InitializeFileLoggingMocked())
         monkeypatch.setattr(toolopts, "CLI", CLIMocked())
         monkeypatch.setattr(main, "show_eula", ShowEulaMocked())
         monkeypatch.setattr(breadcrumbs, "print_data_collection", PrintDataCollectionMocked())
@@ -360,7 +315,6 @@ class TestRollbackFromMain:
 
         assert main.main() == 1
         assert utils.require_root.call_count == 1
-        assert main.initialize_logger.call_count == 1
         assert toolopts.CLI.call_count == 1
         assert main.show_eula.call_count == 1
         assert breadcrumbs.print_data_collection.call_count == 1
@@ -377,8 +331,6 @@ class TestRollbackFromMain:
 
     def test_main_traceback_before_action_completion(self, monkeypatch, caplog, tmp_path):
         require_root_mock = mock.Mock()
-        initialize_logger_mock = mock.Mock()
-        initialize_file_logging_mock = mock.Mock()
         toolopts_cli_mock = mock.Mock()
         show_eula_mock = mock.Mock()
         print_data_collection_mock = mock.Mock()
@@ -397,8 +349,6 @@ class TestRollbackFromMain:
 
         monkeypatch.setattr(applock, "_DEFAULT_LOCK_DIR", str(tmp_path))
         monkeypatch.setattr(utils, "require_root", require_root_mock)
-        monkeypatch.setattr(main, "initialize_logger", initialize_logger_mock)
-        monkeypatch.setattr(main, "initialize_file_logging", initialize_file_logging_mock)
         monkeypatch.setattr(toolopts, "CLI", toolopts_cli_mock)
         monkeypatch.setattr(main, "show_eula", show_eula_mock)
         monkeypatch.setattr(breadcrumbs, "print_data_collection", print_data_collection_mock)
@@ -414,9 +364,7 @@ class TestRollbackFromMain:
         monkeypatch.setattr(report, "summary_as_txt", summary_as_txt_mock)
 
         assert main.main() == 1
-        assert initialize_logger_mock.call_count == 1
         assert require_root_mock.call_count == 1
-        assert initialize_file_logging_mock.call_count == 1
         assert toolopts_cli_mock.call_count == 1
         assert show_eula_mock.call_count == 1
         assert print_data_collection_mock.call_count == 1
@@ -437,8 +385,6 @@ class TestRollbackFromMain:
 
     def test_main_rollback_pre_ponr_changes_phase(self, monkeypatch, caplog, tmp_path):
         require_root_mock = mock.Mock()
-        initialize_logger_mock = mock.Mock()
-        initialize_file_logging_mock = mock.Mock()
         toolopts_cli_mock = mock.Mock()
         show_eula_mock = mock.Mock()
         print_data_collection_mock = mock.Mock()
@@ -459,8 +405,6 @@ class TestRollbackFromMain:
 
         monkeypatch.setattr(applock, "_DEFAULT_LOCK_DIR", str(tmp_path))
         monkeypatch.setattr(utils, "require_root", require_root_mock)
-        monkeypatch.setattr(main, "initialize_logger", initialize_logger_mock)
-        monkeypatch.setattr(main, "initialize_file_logging", initialize_file_logging_mock)
         monkeypatch.setattr(toolopts, "CLI", toolopts_cli_mock)
         monkeypatch.setattr(main, "show_eula", show_eula_mock)
         monkeypatch.setattr(breadcrumbs, "print_data_collection", print_data_collection_mock)
@@ -478,9 +422,7 @@ class TestRollbackFromMain:
         monkeypatch.setattr(report, "summary_as_txt", summary_as_txt_mock)
 
         assert main.main() == 1
-        assert initialize_logger_mock.call_count == 1
         assert require_root_mock.call_count == 1
-        assert initialize_file_logging_mock.call_count == 1
         assert toolopts_cli_mock.call_count == 1
         assert show_eula_mock.call_count == 1
         assert print_data_collection_mock.call_count == 1
@@ -500,8 +442,6 @@ class TestRollbackFromMain:
 
     def test_main_rollback_analyze_exit_phase(self, global_tool_opts, monkeypatch, tmp_path):
         require_root_mock = mock.Mock()
-        initialize_logger_mock = mock.Mock()
-        initialize_file_logging_mock = mock.Mock()
         toolopts_cli_mock = mock.Mock()
         show_eula_mock = mock.Mock()
         print_data_collection_mock = mock.Mock()
@@ -521,8 +461,6 @@ class TestRollbackFromMain:
 
         monkeypatch.setattr(applock, "_DEFAULT_LOCK_DIR", str(tmp_path))
         monkeypatch.setattr(utils, "require_root", require_root_mock)
-        monkeypatch.setattr(main, "initialize_logger", initialize_logger_mock)
-        monkeypatch.setattr(main, "initialize_file_logging", initialize_file_logging_mock)
         monkeypatch.setattr(toolopts, "CLI", toolopts_cli_mock)
         monkeypatch.setattr(main, "show_eula", show_eula_mock)
         monkeypatch.setattr(breadcrumbs, "print_data_collection", print_data_collection_mock)
@@ -542,8 +480,6 @@ class TestRollbackFromMain:
 
         assert main.main() == 0
         assert require_root_mock.call_count == 1
-        assert initialize_logger_mock.call_count == 1
-        assert initialize_file_logging_mock.call_count == 1
         assert toolopts_cli_mock.call_count == 1
         assert show_eula_mock.call_count == 1
         assert print_data_collection_mock.call_count == 1
@@ -560,8 +496,6 @@ class TestRollbackFromMain:
 
     def test_main_rollback_post_ponr_changes_phase(self, monkeypatch, caplog, tmp_path):
         require_root_mock = mock.Mock()
-        initialize_logger_mock = mock.Mock()
-        initialize_file_logging_mock = mock.Mock()
         toolopts_cli_mock = mock.Mock()
         show_eula_mock = mock.Mock()
         print_data_collection_mock = mock.Mock()
@@ -584,8 +518,6 @@ class TestRollbackFromMain:
 
         monkeypatch.setattr(applock, "_DEFAULT_LOCK_DIR", str(tmp_path))
         monkeypatch.setattr(utils, "require_root", require_root_mock)
-        monkeypatch.setattr(main, "initialize_logger", initialize_logger_mock)
-        monkeypatch.setattr(main, "initialize_file_logging", initialize_file_logging_mock)
         monkeypatch.setattr(toolopts, "CLI", toolopts_cli_mock)
         monkeypatch.setattr(main, "show_eula", show_eula_mock)
         monkeypatch.setattr(breadcrumbs, "print_data_collection", print_data_collection_mock)
@@ -606,8 +538,6 @@ class TestRollbackFromMain:
 
         assert main.main() == 1
         assert require_root_mock.call_count == 1
-        assert initialize_logger_mock.call_count == 1
-        assert initialize_file_logging_mock.call_count == 1
         assert toolopts_cli_mock.call_count == 1
         assert show_eula_mock.call_count == 1
         assert print_data_collection_mock.call_count == 1
